+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to the "cli-wrapper" will be documented in this file.
 ### Changed
 
 - Use pnpm scripts instead of npm scripts in generated `package.json`.
+- Use flat EsLint configuration in this project.
+- Update packages to latest versions.
 - Update 'nest' generator to use `example` module.
 - Update 'nest' generator to generate `frontendUrl` env variable.
 - Implement 'app-react' generator.
@@ -14,7 +16,6 @@ All notable changes to the "cli-wrapper" will be documented in this file.
 - Change inputs to only have one level of project type, instead of project/template type combination that was previsouly used.
 - Update `package.json` in publishable projects to include packaging scripts and other smaller changes.
 - Update tsconfig lib to use `es2023` instead of `es2022`.
-- Update packages to latest versions.
 - Remove references to `@gmjs/tsconfig` and `@gmjs/prettier-config` from generated code and the project, and use direct configurations instead.
 - Other generated files improvements and fixes.